Add X-Expires-After header, clean up unit tests
Rather than delete each object from the cache after each request, we
decided that it would be better to remove it from the cache after a
given time interval; this allows the cached objects to be used in
Discord attachments.

Support for a custom self-destruct was added with the X-Expires-After
header, should this functionality be desired.

Unit Tests updated to reflect these changes.